+++ app/src/main/java/com/konkuk/medicarecall/data/repositoryimpl/HomeRepositoryImpl.kt
@@ -14,7 +14,7 @@ class HomeRepositoryImpl(
     private val homeService: HomeService,
 ) : HomeRepository {
     override suspend fun requestImmediateCareCall(
-        elderId: Int,
+        elderId: Long,
         careCallOption: String,
     ): Result<Unit> = runCatching {
         homeService.requestImmediateCareCall(
@@ -22,7 +22,7 @@ class HomeRepositoryImpl(
         ).handleNullableResponse()
     }
 
-    override suspend fun getHomeSummary(elderId: Int): Result<Home> = runCatching {
+    override suspend fun getHomeSummary(elderId: Long): Result<Home> = runCatching {
         homeService.getHomeSummary(elderId).handleResponse().toHome()
     }
 }
